Lets use intent to communicate with systemUI in future

+++ app/src/main/java/com/example/myapplication/data/Sys.kt
@@ -9,9 +9,9 @@ data class Sys(
     @SerializedName("id")
     val id: Int,
     @SerializedName("sunrise")
-    val sunrise: Int,
+    val sunrise: Long,
     @SerializedName("sunset")
-    val sunset: Int,
+    val sunset: Long,
     @SerializedName("type")
     val type: Int
 )
